requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyJWT[crypto]==2.10.1
